refactor: UpdateEmployeeRequest.java renamed to EmployeeDto field injection transformed constructor injection Replaced multiple request parameters with RequestBody EmployeeDto for better readability and maintainability. SpringRestApiAPplication.java moved to com.sever redundant starter package removed
