+++ api-dnm/src/main/java/com/server/SpringRestApiApplication.java
@@ -1,10 +1,10 @@
-package com.server.starter;
+package com.server;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.ComponentScan;
 
-@ComponentScan(basePackages = {"com.server"})
+
+
 @SpringBootApplication
 public class SpringRestApiApplication {
     public static void main(String[] args){
