+++ api-dnm/src/main/java/com/server/services/EmployeeService.java
@@ -1,9 +1,8 @@
 package com.server.services;
 
+import com.server.dto.EmployeeDto;
 import com.server.model.Employee;
-import com.server.model.UpdateEmployeeRequest;
 import com.server.repository.EmployeeRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -11,8 +10,10 @@ import java.util.List;
 @Service
 public class EmployeeService {
 
-    @Autowired
-    private EmployeeRepository employeeRepository;
+    public EmployeeService(EmployeeRepository employeeRepository){
+        this.employeeRepository = employeeRepository;
+    }
+    private final EmployeeRepository employeeRepository;
 
     public List<Employee> getAllEmployeeList(){
         return employeeRepository.getAllEmployeeList();
@@ -22,19 +23,21 @@ public class EmployeeService {
         return employeeRepository.getEmployeeById(id);
     }
 
-    public List<Employee> getEmployeeWithParams(String firstname, String lastname){
-        return employeeRepository.getEmployeeWithParams(firstname,lastname);
+    public List<Employee> getEmployeeWithParams(EmployeeDto employeeDto){
+        return employeeRepository.getEmployeeWithParams(employeeDto.getFirstname(),employeeDto.getLastname());
     }
 
     public Employee saveEmployee(Employee newEmployee){
+
         return employeeRepository.saveEmployee(newEmployee);
     }
 
     public boolean deleteEmployee(String id){
+
         return employeeRepository.deleteEmployee(id);
     }
 
-    public Employee updateEmployee(String id, UpdateEmployeeRequest request){
+    public Employee updateEmployee(String id, EmployeeDto request){
         return employeeRepository.updateEmployee(id,request);
     }
 }
